requirements.txt
appdirs==1.4.4
astroid==2.4.2
attrs==19.3.0
black==19.10b0
click==7.1.2
colorama==0.4.3
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathspec==0.8.0
PyJWT==1.7.1
pylint==2.5.3
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
